libpmc: rapl_allocate_pmc: drop unneeded enum casts
Requested in D58029 review.  PMC_EV_RAPL_FIRST/LAST are members of enum
pmc_event, so pe compares against them directly; the (int) casts were
copied from soft_allocate_pmc, where they are needed because the SOFT
bounds are out-of-enum #defines for the dynamic software-event range.
This also matches the event-name lookup, which already compares pe to
the RAPL bounds without a cast.

+++ lib/libpmc/libpmc.c
@@ -841,7 +841,7 @@ static int
 rapl_allocate_pmc(enum pmc_event pe, char *ctrspec,
     struct pmc_op_pmcallocate *pmc_config)
 {
-	if ((int)pe < PMC_EV_RAPL_FIRST || (int)pe > PMC_EV_RAPL_LAST)
+	if (pe < PMC_EV_RAPL_FIRST || pe > PMC_EV_RAPL_LAST)
 		return (-1);
 
 	/* RAPL events must be unqualified. */
